requirements.txt
faiss-cpu==1.6.1
nmslib==2.0.5
jieba>=0.39
joblib>=0.13.2
numpy>=1.15.4
pandas>=0.23.4
scikit-learn>=0.22.1
zhon==1.1.5